Updated to add openssl-1.1 and Steam as dependencies, and a script to handle missing $HOME folders

+++ PKGBUILD
@@ -7,12 +7,12 @@
 # If you want additional options, there are switches below.
 pkgname=unreal-engine
 pkgver=5.0.3
-pkgrel=0
+pkgrel=1
 pkgdesc='A 3D game engine by Epic Games which can be used non-commercially for free.'
 arch=('x86_64' 'x86_64_v2' 'x86_64_v3' 'x86_64_v4' 'aarch64')
 url=https://www.unrealengine.com/
-makedepends=('git' 'openssh' 'coreutils' 'sed' 'grep' 'glibc')
-depends=('icu63' 'sdl2' 'python' 'lld' 'xdg-user-dirs' 'dos2unix')
+makedepends=('git' 'openssh' 'sed' 'grep' 'glibc')
+depends=('icu63' 'sdl2' 'python' 'lld' 'xdg-user-dirs' 'dos2unix' 'openssl-1.1' 'steam' 'coreutils' 'findutils')
 optdepends=('qt5-base: qmake build system for projects'
             'cmake: build system for projects'
             'qtcreator: IDE for projects'
@@ -21,21 +21,37 @@ optdepends=('qt5-base: qmake build system for projects'
             'clion: IDE for projects'
             'pacman-contrib: for the paccache cleaning hook')
 license=('custom:UnrealEngine')
-source=('com.unrealengine.UE4Editor.desktop'
+source=('unreal-engine-5.sh'
+        'com.unrealengine.UE4Editor.desktop'
         'use_system_clang.patch'
-        'unreal-engine-pacman-cache.hook')
-sha256sums=('15e9f9d8dc8bd8513f6a5eca990e2aab21fd38724ad57d213b06a6610a951d58'
+        'unreal-engine-5-pacman-cache.hook')
+sha256sums=('9358ec38944f712101d733eb7b4cfa082b47a5021e42809850781c462a826644'
+            'c04c03b2c5c933b7eb1af283d607934ad95fd57f44d62b83719061b555a85dca'
             'b0a57db9a44d0001dc76ca8504d93e273af30093c6a993a5969d82b0ace54b98'
             '9386160a91594abeeaf4fe02fea562e7a4ead4c6f9a258c2a37b2e5f10e7deca')
 # Not sure if compiling Unreal with LTO is legal? Lot's of different proprietary software goes into Unreal
-options=('!strip' 'staticlibs') # Package is 3 Gib smaller with "strip" but it takes a long time and generates many warnings
+options=('!strip' 'staticlibs') # Package is smaller with "strip" but it takes a long time and generates many warnings
 
 ## Use this if you prefer opendoas; the benefit here is that doas won't time out on you if you wait too long to authenticate after compilation
 if [ -f /usr/bin/doas ] && [ -f /etc/doas.conf ]; then
   PACMAN_AUTH=(doas)
 fi
 
-_use_system_clang=false # Change this to true for a potentially smaller package size
+# Change this to true or 1 for a potentially smaller package size
+## Note: We can't guarantee that enabling this won't affect cross-distro compatibility given that Epic provides their own toolchain specifically
+## based on the claim that it helps with compatibility...
+## Though, you could just distribute your game as an AppImage if you decide to publish on a game store... :p
+_use_system_clang=false
+
+# Default engine installation directory. Can be useful if you do not have a lot of space on the default storage drive
+# DON'T put a "/" at the start of the path
+_install_dir="opt/${pkgname}"
+
+## An attempt to fix the NuGet SSL issue during the build -- didn't seem to work; users had to do this manually, so we'll rollback to an alternative
+#ln -s /etc/ssl /usr/lib/ssl
+#ln -s /etc/ssl /usr/lib64/ssl
+## ---
+export DOTNET_SYSTEM_NET_HTTP_USESOCKETSHTTPHANDLER=0
 
 ## This is for detecting your CPU architecture automatically; set to false if you want to enforce your own makepkg.conf file
 ## Disabled by default as a compromise for those bothered by having it force-enabled
@@ -52,24 +68,24 @@ then
   ## Architecture checks and compile flag adjustments - shellcheck throws a fit about the build function but it looks fine to me; checks for the highest available x64 support level and falls back to "native" if either not available
   if [ "$(uname -m)" == "x86_64" ]; then
     if [ "$(/lib/ld-linux-x86-64.so.2 --help | grep -w 'x86-64-v4' | cut -d ',' -f 1 | sed 's/^  //' | sed 's/ (/ - /')" == '  x86-64-v4 - supported' ]; then
-      export CFLAGS="${CFLAGS} -march=x86-64-v4 -mtune=x86-64-v4 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+      export CFLAGS="${CFLAGS} -march=x86-64-v4 -mtune=x86-64-v4 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
       export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
       export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
     elif [ "$(/lib/ld-linux-x86-64.so.2 --help | grep -w 'x86-64-v3' | cut -d ',' -f 1 | sed 's/^  //' | sed 's/ (/ - /')" == 'x86-64-v3 - supported' ]; then
-      export CFLAGS="${CFLAGS} -march=x86-64-v3 -mtune=x86-64-v3 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+      export CFLAGS="${CFLAGS} -march=x86-64-v3 -mtune=x86-64-v3 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
       export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
       export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
     elif [ "$(/lib/ld-linux-x86-64.so.2 --help | grep -w 'x86-64-v2' | cut -d ',' -f 1 | sed 's/^  //' | sed 's/ (/ - /')" == 'x86-64-v2 - supported' ]; then
-      export CFLAGS="${CFLAGS} -march=x86-64-v2 -mtune=x86-64-v2 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+      export CFLAGS="${CFLAGS} -march=x86-64-v2 -mtune=x86-64-v2 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
       export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
       export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
     elif [ "$(/lib/ld-linux-x86-64.so.2 --help | grep 'x86_64' | grep 'supported' | cut -d ',' -f 1 | sed 's/^  //' | sed 's/ (/ - /' | grep -w '^x86_64 - supported')" == 'x86_64 - supported' ]; then
-      export CFLAGS="${CFLAGS} -march=x86-64 -mtune=x86-64 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+      export CFLAGS="${CFLAGS} -march=x86-64 -mtune=x86-64 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
       export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
       export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
     fi
   elif [ "$(uname -m)" == "aarch64" ]; then
-    export CFLAGS="${CFLAGS} -march=aarch64 -mtune=aarch64 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+    export CFLAGS="${CFLAGS} -march=aarch64 -mtune=aarch64 -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
     export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
     export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
   else
@@ -77,20 +93,11 @@ then
     return
   fi
 elif [[ "${arch_auto}" == native ]]; then
-    export CFLAGS="${CFLAGS} -march=native -mtune=native -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -fvisibility=hidden -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
+    export CFLAGS="${CFLAGS} -march=native -mtune=native -O3 -pipe -fno-plt -fstack-clash-protection -fstack-protector-strong -fcf-protection -fsanitize=bounds,alignment,object-size -fsanitize-undefined-trap-on-error -Wformat -Werror=format-security -Warray-bounds -Wvla -Wimplicit-fallthrough -Wno-unused-result -Wno-unneeded-internal-declaration"
     export CXXFLAGS="${CFLAGS} -Wp,-D_GLIBCXX_ASSERTIONS"
     export LDFLAGS="-Wl,-O3,--sort-common,--as-needed,-z,relro,-z,now"
 fi
 
-## An attempt to fix the NuGet SSL issue during the build -- didn't seem to work; users had to do this manually, so we'll rollback to an alternative
-#ln -s /etc/ssl /usr/lib/ssl
-#ln -s /etc/ssl /usr/lib64/ssl
-## ---
-export DOTNET_SYSTEM_NET_HTTP_USESOCKETSHTTPHANDLER=0
-
-# Default engine installation directory. Can be useful if you do not have a lot of space on the default storage drive
-# DON'T put a "/" at the start of the path
-_install_dir="opt/${pkgname}"
 
 prepare() {
   # Check access to the repository
@@ -101,20 +108,20 @@ prepare() {
   fi
 
   # Download Unreal Engine source or update if the folder exists
-  if [[ ! -d ${pkgname} ]]
+  if [[ ! -d "${pkgname}" ]]
   then
-    #git clone --depth=1 --branch=ue5-main git@github.com:EpicGames/UnrealEngine {$pkgname}
-    git clone --depth=1 --branch=${pkgver}-release git@github.com:EpicGames/UnrealEngine ${pkgname}
-    cd ${pkgname} || return
+    #git clone --depth=1 --branch=ue5-main git@github.com:EpicGames/UnrealEngine "{$pkgname}"
+    git clone --depth=1 --branch=${pkgver}-release git@github.com:EpicGames/UnrealEngine "${pkgname}"
+    cd "${pkgname}" || return
   else
-    cd ${pkgname} || return
+    cd "${pkgname}" || return
     rm -f .git/index.lock
     git fetch --depth=1 origin tag @{upstream}
     git reset --hard @{upstream}
   fi
   
   ## Apply custom patches if enabled
-  if [[ ${_use_system_clang} == true ]]; then
+  if [ "${_use_system_clang}" == true ] || [ "${_use_system_clang}" -eq 1 ]; then
     patch -p1 -i "${srcdir}/use_system_clang.patch"
   fi
 
@@ -136,43 +143,28 @@ prepare() {
 }
 
 build() {
-  cd ${pkgname}
+  cd "${pkgname}" || return
   Engine/Build/BatchFiles/RunUAT.sh BuildGraph -target="Make Installed Build Linux" -script=Engine/Build/InstalledEngineBuild.xml -set:WithDDC=false -set:HostPlatformOnly=false -set:WithLinux=true -set:WithWin64=true -set:WithWin32=false -set:WithMac=false -set:WithAndroid=false -set:WithIOS=false -set:WithTVOS=false -set:WithLumin=false -set:WithHTML5=false
 }
 
 package() {
 
-# Desktop entry
-if [[ -f com.unrealengine.UE4Editor.desktop ]] || [[ -f com.unrealengine.UE5Editor.desktop ]] && [[ "${_install_dir}" != "opt/${pkgname}" ]] 
-then
-  sed -i "5c\Path=/${_install_dir}/Engine/Binaries/Linux/" com.unrealengine.UE4Editor.desktop
-  sed -i "6c\Exec=/${_install_dir}/Engine/Binaries/Linux/UnrealEditor %F" com.unrealengine.UE4Editor.desktop
-  sed -i 's/Unreal Engine 4 Editor/Unreal Engine 5 Editor/g' com.unrealengine.UE4Editor.desktop
-  sed -i 's/Icon=ue4editor/Icon=ue5editor/g' com.unrealengine.UE4Editor.desktop
-  if [[ -f com.unrealengine.UE4Editor.desktop ]]
-  then
+  # Desktop entry
+  if [ ! -f com.unrealengine.UE5Editor.desktop ] && [ -f com.unrealengine.UE4Editor.desktop ]; then
     cp com.unrealengine.UE4Editor.desktop com.unrealengine.UE5Editor.desktop
   fi
+  
+  sed -i "7c\Exec=/usr/bin/unreal-engine-5.sh %U" com.unrealengine.UE5Editor.desktop
+  sed -i "14c\Path=/usr/bin/" com.unrealengine.UE5Editor.desktop
+  sed -i 's/Unreal Engine 4 Editor/Unreal Engine 5 Editor/g' com.unrealengine.UE5Editor.desktop
+  sed -i 's/Icon=ue4editor/Icon=ue5editor/g' com.unrealengine.UE5Editor.desktop
   install -Dm775 com.unrealengine.UE5Editor.desktop "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop"
-else
-  if [[ -f com.unrealengine.UE4Editor.desktop ]] || [[ -f com.unrealengine.UE5Editor.desktop ]] && [[ "${_install_dir}" == "opt/${pkgname}" ]]
-  then
-    sed -i "5c\Path=/opt/${pkgname}/Engine/Binaries/Linux/" com.unrealengine.UE4Editor.desktop
-    sed -i "6c\Exec=/opt/${pkgname}/Engine/Binaries/Linux/UnrealEditor %F" com.unrealengine.UE4Editor.desktop
-    sed -i 's/Unreal Engine 4 Editor/Unreal Engine 5 Editor/g' com.unrealengine.UE4Editor.desktop
-    sed -i 's/Icon=ue4editor/Icon=ue5editor/g' com.unrealengine.UE4Editor.desktop
-    if [[ -f com.unrealengine.UE4Editor.desktop ]]
-    then
-      cp com.unrealengine.UE4Editor.desktop com.unrealengine.UE5Editor.desktop
-    fi
-    install -Dm775 com.unrealengine.UE5Editor.desktop "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop"
-  fi
-fi
+  chmod +x "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop"
 
   ## Install a pacman hook to keep old builds from compounding cache by tens of GBs - 2 builds alone can reach at least 30 GBs in pacman's cache; having one only takes up about 15 GBs
-  install -Dm775 unreal-engine-pacman-cache.hook "${pkgdir}/etc/pacman.d/hooks/unreal-engine-pacman-cache.hook"
+  install -Dm775 unreal-engine-5-pacman-cache.hook "${pkgdir}/etc/pacman.d/hooks/unreal-engine-5-pacman-cache.hook"
   
-  cd ${pkgname} || return
+  cd "${pkgname}" || return
   
   # Icon for Desktop entry
   install -Dm770 Engine/Source/Programs/UnrealVS/UnrealVS.2022/Resources/Preview.png "${pkgdir}/usr/share/pixmaps/ue5editor.png"
@@ -183,6 +175,24 @@ fi
   # Engine
   ## Set to all permissions to prevent the engine from breaking itself; more elegant solutions might exist - suggest them if they can be automated here
   ## Also, correct me if I package this improperly; I added Win64 support for the build in hopes of supporting cross-compilation
-  install -dm777 "${pkgdir}/${_install_dir}/Engine"
-  mv LocalBuilds/Engine/Linux/* "${pkgdir}/${_install_dir}"
+  install -Dm777 "${pkgdir}/${_install_dir}/Engine"
+  install -Dm777 LocalBuilds/Engine/Linux/ "${pkgdir}/${_install_dir}/LocalBuilds/Engine/Linux/"
+  
+  chmod +x "${pkgdir}/${_install_dir}/Engine/Binaries/ThirdParty/Mono/Linux/bin/xbuild"
+  chmod +x "${pkgdir}/${_install_dir}/Engine/Binaries/ThirdParty/Mono/Linux/bin/mcs"
+  mkdir -p "${pkgdir}/${_install_dir}/Engine/Binaries/Android/"
+  
+  # Launch script to initialize missing user folders for Unreal Engine
+  install -Dm755 unreal-engine-5.sh "${pkgdir}/usr/bin/"
+  chmod +x "${pkgdir}/usr/bin/unreal-engine-5.sh"
+  ln -s "${pkgdir}/usr/bin/unreal-engine-5.sh" "${pkgdir}/usr/bin/ue5"
+  ln -s "${pkgdir}/usr/bin/unreal-engine-5.sh" "${pkgdir}/usr/bin/UE5"
+  ln -s "${pkgdir}/usr/bin/unreal-engine-5.sh" "${pkgdir}/usr/bin/unreal-engine-5"
+  chmod 755 "${pkgdir}/usr/bin/ue5" "${pkgdir}/usr/bin/UE5" "${pkgdir}/usr/bin/unreal-engine-5"
+  
+  # Configuring the launch script to detect when it has been run for the first time
+  # Note: Requires that there isn't already a UE5 desktop entry in "${HOME}/local/share/applications/" - delete yours if you have one there before installing this
+  DesktopFileChecksum=$(sha256sum "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop" | cut -f 1 -d ' ')
+  sed -i "s/ChecksumPlaceholder/${DesktopFileChecksum}/" "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop"
+  sed -i "s/InstalledLocationPlaceholder/${_install_dir}/" "${pkgdir}/usr/share/applications/com.unrealengine.UE5Editor.desktop"
 }
